+++ core/src/main/java/com/expediagroup/streamplatform/streamregistry/core/services/ProducerService.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2018-2021 Expedia, Inc.
+ * Copyright (C) 2018-2022 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -103,15 +103,4 @@ public class ProducerService {
   public Producer canCreateProducer(Producer producer) {
     return producer;
   }
-
-  @PreAuthorize("hasPermission(#producer, 'UPDATE')")
-  public Producer canUpdateProducer(Producer producer) {
-    return producer;
-  }
-
-  @PreAuthorize("hasPermission(#producer, 'DELETE')")
-  public Producer canDeleteProducer(Producer producer) {
-    return producer;
-  }
-
 }