+++ core/src/main/java/com/expediagroup/streamplatform/streamregistry/core/validators/ConsumerValidator.java
@@ -20,6 +20,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
 import com.expediagroup.streamplatform.streamregistry.core.validators.key.KeyValidator;
+import com.expediagroup.streamplatform.streamregistry.core.views.DomainView;
 import com.expediagroup.streamplatform.streamregistry.core.views.StreamView;
 import com.expediagroup.streamplatform.streamregistry.core.views.ZoneView;
 import com.expediagroup.streamplatform.streamregistry.model.Consumer;
@@ -29,6 +30,7 @@ import com.expediagroup.streamplatform.streamregistry.model.Consumer;
 public class ConsumerValidator implements Validator<Consumer> {
   private final StreamView streamView;
   private final ZoneView zoneView;
+  private final DomainView domainView;
   private final KeyValidator<Consumer> consumerKeyValidator;
   private final SpecificationValidator specificationValidator;
 
@@ -46,6 +48,10 @@ public class ConsumerValidator implements Validator<Consumer> {
   }
 
   public void validateForCreateAndUpdate(Consumer consumer) throws ValidationException {
+    if (!domainView.exists(consumer.getKey().getStreamKey().getDomainKey())) {
+      throw new ValidationException("Domain does not exist");
+    }
+
     if (!streamView.exists(consumer.getKey().getStreamKey())) {
       throw new ValidationException("Stream does not exist");
     }