+++ core/src/main/java/com/expediagroup/streamplatform/streamregistry/core/validators/ProducerValidator.java
@@ -20,6 +20,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
 import com.expediagroup.streamplatform.streamregistry.core.validators.key.KeyValidator;
+import com.expediagroup.streamplatform.streamregistry.core.views.DomainView;
 import com.expediagroup.streamplatform.streamregistry.core.views.StreamView;
 import com.expediagroup.streamplatform.streamregistry.core.views.ZoneView;
 import com.expediagroup.streamplatform.streamregistry.model.Producer;
@@ -29,6 +30,8 @@ import com.expediagroup.streamplatform.streamregistry.model.Producer;
 public class ProducerValidator implements Validator<Producer> {
   private final StreamView streamView;
   private final ZoneView zoneView;
+
+  private final DomainView domainView;
   private final KeyValidator<Producer> producerKeyValidator;
   private final SpecificationValidator specificationValidator;
 
@@ -46,9 +49,14 @@ public class ProducerValidator implements Validator<Producer> {
   }
 
   public void validateForCreateAndUpdate(Producer producer) throws ValidationException {
+    if (!domainView.exists(producer.getKey().getStreamKey().getDomainKey())) {
+      throw new ValidationException("Domain does not exist");
+    }
+
     if (!streamView.exists(producer.getKey().getStreamKey())) {
       throw new ValidationException("Stream does not exist");
     }
+
     if (!zoneView.exists(producer.getKey().getZoneKey())) {
       throw new ValidationException("Zone does not exist");
     }