+++ core/src/test/java/com/expediagroup/streamplatform/streamregistry/core/services/ProcessServiceTest.java
@@ -20,9 +20,11 @@ import static java.util.Collections.emptyList;
 import static java.util.Optional.empty;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doNothing;
+import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.inOrder;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
@@ -119,6 +121,14 @@ public class ProcessServiceTest {
     verify(processRepository).save(entity);
   }
 
+  @Test(expected = ValidationException.class)
+  public void createFailedWhenDomainNotExists() {
+    Process entity = createTestProcess();
+    when(processRepository.findById(entity.getKey())).thenReturn(empty());
+    doThrow(ValidationException.class).when(processValidator).validateForCreate(entity);
+    processService.create(entity);
+  }
+
   @Test(expected = AccessDeniedException.class)
   public void createFailAuthConsumer() {
     Process p = createTestProcess();
@@ -134,7 +144,7 @@ public class ProcessServiceTest {
   }
 
   @Test
-  public void update() {
+  public void updateNoInputOutputChanges() {
     final Specification specification = mock(Specification.class);
 
     Process existing = createTestProcess();
@@ -150,13 +160,13 @@ public class ProcessServiceTest {
     verify(processRepository).findById(existing.getKey());
     verify(processValidator).validateForUpdate(newEntity, existing);
     verify(handlerService).handleUpdate(newEntity, existing);
-    verify(consumerService).canUpdateConsumer(any(Consumer.class));
-    verify(producerService).canUpdateProducer(any(Producer.class));
+    verify(consumerService, never()).canCreateConsumer(any(Consumer.class));
+    verify(producerService, never()).canCreateProducer(any(Producer.class));
     verify(processRepository).save(newEntity);
   }
 
   @Test
-  public void updateCreate() {
+  public void updateWithInputOutputChanges() {
     final Specification specification = mock(Specification.class);
 
     Process existing = createTestProcess();
@@ -176,34 +186,8 @@ public class ProcessServiceTest {
     verify(processRepository).findById(existing.getKey());
     verify(processValidator).validateForUpdate(newEntity, existing);
     verify(handlerService).handleUpdate(newEntity, existing);
-    verify(consumerService).canCreateConsumer(any(Consumer.class));
-    verify(producerService).canCreateProducer(any(Producer.class));
-    verify(processRepository).save(newEntity);
-  }
-
-  @Test
-  public void updateDelete() {
-    final Specification specification = mock(Specification.class);
-
-    Process existing = createTestProcess();
-    Process newEntity = createTestProcess();
-    ProcessInputStream pis = new ProcessInputStream(new StreamKey("inputDomain","streamInputName2",1), new ObjectMapper().createObjectNode());
-    existing.getInputs().add(pis);
-    ProcessOutputStream pos = new ProcessOutputStream(new StreamKey("outputDomain","streamOutputName2",1), new ObjectMapper().createObjectNode());
-    existing.getOutputs().add(pos);
-
-    when(processRepository.findById(existing.getKey())).thenReturn(Optional.of(existing));
-    doNothing().when(processValidator).validateForUpdate(newEntity, existing);
-    when(handlerService.handleUpdate(newEntity, existing)).thenReturn(specification);
-    when(processRepository.save(newEntity)).thenReturn(newEntity);
-
-    processService.update(newEntity);
-
-    verify(processRepository).findById(existing.getKey());
-    verify(processValidator).validateForUpdate(newEntity, existing);
-    verify(handlerService).handleUpdate(newEntity, existing);
-    verify(consumerService).canDeleteConsumer(any(Consumer.class));
-    verify(producerService).canDeleteProducer(any(Producer.class));
+    verify(consumerService, times(1)).canCreateConsumer(any(Consumer.class));
+    verify(producerService, times(1)).canCreateProducer(any(Producer.class));
     verify(processRepository).save(newEntity);
   }
 
@@ -213,24 +197,48 @@ public class ProcessServiceTest {
     processService.update(p);
   }
 
+  @Test(expected = ValidationException.class)
+  public void updateFailedWhenDomainNotExists() {
+    Process entity = createTestProcess();
+    Process existing = createTestProcess();
+    when(processRepository.findById(existing.getKey())).thenReturn(Optional.of(existing));
+    doThrow(ValidationException.class).when(processValidator).validateForUpdate(entity, existing);
+    processService.update(entity);
+  }
+
   @Test(expected = AccessDeniedException.class)
   public void updateFailAuthConsumer() {
-    Process p = createTestProcess();
+    final Specification specification = mock(Specification.class);
 
-    when(processRepository.findById(p.getKey())).thenReturn(Optional.of(p));
-    when(consumerService.canUpdateConsumer(any(Consumer.class))).thenThrow(AccessDeniedException.class);
+    Process existing = createTestProcess();
+    Process newEntity = createTestProcess();
 
-    processService.update(p);
+    ProcessInputStream pis = new ProcessInputStream(new StreamKey("inputDomain","streamInputName2",1), new ObjectMapper().createObjectNode());
+    newEntity.getInputs().add(pis);
+
+    when(processRepository.findById(existing.getKey())).thenReturn(Optional.of(existing));
+    doNothing().when(processValidator).validateForUpdate(newEntity, existing);
+    when(handlerService.handleUpdate(newEntity, existing)).thenReturn(specification);
+    when(consumerService.canCreateConsumer(any(Consumer.class))).thenThrow(AccessDeniedException.class);
+
+    processService.update(newEntity);
   }
 
   @Test(expected = AccessDeniedException.class)
   public void updateFailAuthProducer() {
-    Process p = createTestProcess();
+    final Specification specification = mock(Specification.class);
 
-    when(processRepository.findById(p.getKey())).thenReturn(Optional.of(p));
-    when(producerService.canUpdateProducer(any(Producer.class))).thenThrow(AccessDeniedException.class);
+    Process existing = createTestProcess();
+    Process newEntity = createTestProcess();
+    ProcessOutputStream pos = new ProcessOutputStream(new StreamKey("outputDomain","streamOutputName2",1), new ObjectMapper().createObjectNode());
+    newEntity.getOutputs().add(pos);
 
-    processService.update(p);
+    when(processRepository.findById(existing.getKey())).thenReturn(Optional.of(existing));
+    doNothing().when(processValidator).validateForUpdate(newEntity, existing);
+    when(handlerService.handleUpdate(newEntity, existing)).thenReturn(specification);
+    when(producerService.canCreateProducer(any(Producer.class))).thenThrow(AccessDeniedException.class);
+
+    processService.update(newEntity);
   }
 
   @Test
@@ -319,8 +327,7 @@ public class ProcessServiceTest {
     p.setKey(new ProcessKey("domain","name"));
     List<ZoneKey> zones = new ArrayList<>();
     zones.add(new ZoneKey("aws_us_east_1"));
-    p.setZones(zones);
+    p.setZoneKeys(zones);
     return p;
   }
-
 }