+++ model/src/main/java/com/expediagroup/streamplatform/streamregistry/model/Process.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2018-2021 Expedia, Inc.
+ * Copyright (C) 2018-2022 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -30,7 +30,7 @@ import com.expediagroup.streamplatform.streamregistry.model.keys.ZoneKey;
 public class Process implements Entity<ProcessKey> {
   private ProcessKey key;
   private Specification specification;
-  private List<ZoneKey> zones;
+  private List<ZoneKey> zoneKeys;
   private List<ProcessInputStream> inputs;
   private List<ProcessOutputStream> outputs;
   private Status status;