+++ model/src/main/java/com/expediagroup/streamplatform/streamregistry/model/ProcessInputStream.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2018-2021 Expedia, Inc.
+ * Copyright (C) 2018-2022 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,6 +23,6 @@ import com.expediagroup.streamplatform.streamregistry.model.keys.StreamKey;
 
 @Value
 public class ProcessInputStream {
-  StreamKey stream;
+  StreamKey streamKey;
   ObjectNode configuration;
 }